improve camera controls: add cinematic mode, free camera with smooth …

+++ src/core/CelestialBody.cpp
@@ -7,11 +7,11 @@
 CelestialBody::CelestialBody()
     : position(Eigen::Vector3f::Zero()), velocity(Eigen::Vector3f::Zero()), force(Eigen::Vector3f::Zero()), mass(1.0f), radius(1.0f) {}
 
-CelestialBody::CelestialBody(const Eigen::Vector3f& pos, const Eigen::Vector3f& vel, float m)
+CelestialBody::CelestialBody(const Eigen::Vector3f& pos, const Eigen::Vector3f& vel, double m)
     : position(pos), velocity(vel), force(Eigen::Vector3f::Zero()), mass(m), radius(calculateRadius(m)) {}
 
-float CelestialBody::calculateRadius(float mass) {
-    if (mass <= 0) return 0.0f;
+double CelestialBody::calculateRadius(double mass) {
+    if (mass <= 0) return 0.0;
     return std::cbrt(mass); // Temporary approximation of radius (If we assume that density is constance for every body, then R is proportional to cbrt(m))
 }
 
@@ -28,7 +28,7 @@ void CelestialBody::addForce(Eigen::Vector3f& force) { this->force += force; }
 //     position += velocity * deltaTime;
 // }
 
-void CelestialBody::updatePosition(float deltaTime) {
+void CelestialBody::updatePosition(double deltaTime) {
     Eigen::Vector3f acceleration = force / mass;
 
     // Step 1: Position update
@@ -38,7 +38,7 @@ void CelestialBody::updatePosition(float deltaTime) {
     velocity += 0.5f * acceleration * deltaTime;
 }
 
-void CelestialBody::updateVelocity(float deltaTime) {
+void CelestialBody::updateVelocity(double deltaTime) {
     Eigen::Vector3f newAcceleration = force / mass;
 
     velocity += 0.5f * newAcceleration * deltaTime;
